requirements.txt
click
pillow
ocrd_utils
ocrd_models
